+++ docs/changelog.rst
@@ -12,6 +12,7 @@ Changelog
 
 * Fixed intra-frame button press/release detection.
 * Added ``--hidden`` CLI option.
+* Fixed HashMap and HashSet.
 
 **Tools**